split models into modules
