+++ profiles/services/staff_sso.py
@@ -2,14 +2,8 @@ from typing import TYPE_CHECKING, Optional
 
 from django.contrib.auth import get_user_model
 
-from profiles.models import (
-    EMAIL_TYPE_WORK,
-    EMAIL_TYPES,
-    Email,
-    Profile,
-    StaffSSOProfile,
-    StaffSSOProfileEmail,
-)
+from profiles.models.generic import EMAIL_TYPE_WORK, EMAIL_TYPES, Email, Profile
+from profiles.models.staff_sso import StaffSSOProfile, StaffSSOProfileEmail
 
 
 if TYPE_CHECKING:
